refactor: move poly from common to backend
