+++ halo2_backend/src/lib.rs
@@ -1,10 +1,10 @@
+mod helpers;
 pub mod plonk;
+pub mod poly;
 
 // Internal re-exports
 pub use halo2_common::arithmetic;
 pub use halo2_common::circuit;
-pub use halo2_common::helpers;
 pub use halo2_common::multicore;
-pub use halo2_common::poly;
 pub use halo2_common::transcript;
 pub use halo2_common::SerdeFormat;
