+++ halo2_backend/src/poly/kzg/strategy.rs
