+++ halo2_backend/src/poly/strategy.rs
