+++ halo2_common/src/helpers.rs
@@ -1,4 +1,3 @@
-use crate::poly::Polynomial;
 use halo2_middleware::ff::PrimeField;
 use halo2curves::{serde::SerdeObject, CurveAffine};
 use std::io;
@@ -118,37 +117,4 @@ pub fn unpack(byte: u8, bits: &mut [bool]) {
     for (bit_index, bit) in bits.iter_mut().enumerate() {
         *bit = (byte >> bit_index) & 1 == 1;
     }
-}
-
-/// Reads a vector of polynomials from buffer
-pub fn read_polynomial_vec<R: io::Read, F: SerdePrimeField, B>(
-    reader: &mut R,
-    format: SerdeFormat,
-) -> io::Result<Vec<Polynomial<F, B>>> {
-    let mut len = [0u8; 4];
-    reader.read_exact(&mut len)?;
-    let len = u32::from_be_bytes(len);
-
-    (0..len)
-        .map(|_| Polynomial::<F, B>::read(reader, format))
-        .collect::<io::Result<Vec<_>>>()
-}
-
-/// Writes a slice of polynomials to buffer
-pub fn write_polynomial_slice<W: io::Write, F: SerdePrimeField, B>(
-    slice: &[Polynomial<F, B>],
-    writer: &mut W,
-    format: SerdeFormat,
-) -> io::Result<()> {
-    writer.write_all(&(slice.len() as u32).to_be_bytes())?;
-    for poly in slice.iter() {
-        poly.write(writer, format)?;
-    }
-    Ok(())
-}
-
-/// Gets the total number of bytes of a slice of polynomials, assuming all polynomials are the same length
-pub fn polynomial_slice_byte_length<F: PrimeField, B>(slice: &[Polynomial<F, B>]) -> usize {
-    let field_len = F::default().to_repr().as_ref().len();
-    4 + slice.len() * (4 + field_len * slice.get(0).map(|poly| poly.len()).unwrap_or(0))
 }
