+++ halo2_proofs/src/lib.rs
@@ -42,7 +42,7 @@ pub mod dev {
 /// the committed polynomials at arbitrary points.
 pub mod poly {
     pub use halo2_backend::poly::VerificationStrategy;
-    pub use halo2_common::poly::{commitment, ipa, kzg};
+    pub use halo2_backend::poly::{commitment, ipa, kzg};
     pub use halo2_middleware::poly::Rotation;
 }
 /// This module contains utilities and traits for dealing with Fiat-Shamir
