+++ halo2_proofs/src/plonk/prover.rs
@@ -1,6 +1,6 @@
+use crate::poly::commitment::{CommitmentScheme, Params, Prover};
 use halo2_backend::plonk::{prover::ProverV2, ProvingKey};
 use halo2_common::plonk::{circuit::Circuit, Error};
-use halo2_common::poly::commitment::{CommitmentScheme, Params, Prover};
 use halo2_common::transcript::{EncodedChallenge, TranscriptWrite};
 use halo2_frontend::circuit::{compile_circuit, WitnessCalculator};
 use halo2_middleware::ff::{FromUniformBytes, WithSmallOrderMulGroup};
